backend/requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-huggingface
chromadb
rank_bm25
python-dotenv
sentence-transformers
fastapi
uvicorn
langchain-chroma
langchain-openai
